Authentication and Authorization added

+++ src/actions/actions.ts
@@ -4,6 +4,8 @@ import { sleep } from "@/lib/utils";
 import { revalidatePath } from "next/cache";
 import { petFormSchema, petIdSchema } from "@/lib/validations";
 import { signIn, signOut } from "@/lib/auth";
+import bcrypt from "bcryptjs";
+import { checkAuth, getPetById } from "@/lib/server-utils";
 
 // -- User actions --
 
@@ -20,10 +22,30 @@ export async function logOut() {
   });
 }
 
+export async function signUp(formData: FormData) {
+  //hash password to store in DB
+  const hashedPassword = await bcrypt.hash(
+    formData.get("password") as string,
+    10
+  );
+
+  await prisma.user.create({
+    data: {
+      email: formData.get("email") as string,
+      hashedPassword,
+    },
+  });
+
+  //then using NextAuth signIn method to sign in user which will then redirect to app page
+  await signIn("credentials", formData);
+}
+
 // -- Pet actions --
 
 export async function addPet(pet: unknown) {
   await sleep(1000);
+
+  const session = await checkAuth();
 
   const validatedPet = petFormSchema.safeParse(pet);
 
@@ -35,7 +57,14 @@ export async function addPet(pet: unknown) {
 
   try {
     await prisma.pet.create({
-      data: validatedPet.data,
+      data: {
+        ...validatedPet.data,
+        user: {
+          connect: {
+            id: session.user.id,
+          },
+        },
+      },
     });
   } catch (error) {
     return {
@@ -49,6 +78,10 @@ export async function addPet(pet: unknown) {
 export async function editPet(petId: unknown, newPetData: unknown) {
   await sleep(1000);
 
+  //authentication check -> i.e their token is valid
+  const session = await checkAuth();
+
+  //validation
   const validatedPet = petFormSchema.safeParse(newPetData);
   const validatedPetId = petIdSchema.safeParse(petId);
 
@@ -58,6 +91,26 @@ export async function editPet(petId: unknown, newPetData: unknown) {
     };
   }
 
+  //authorization check -> i.e user is authorized to edit this pet
+  const pet = await prisma.pet.findUnique({
+    where: {
+      id: validatedPetId.data,
+    },
+  });
+
+  if (!pet) {
+    return {
+      message: "Pet not found",
+    };
+  }
+
+  if (pet.userId !== session.user.id) {
+    return {
+      message: "Unauthorized",
+    };
+  }
+
+  //database mutation
   try {
     await prisma.pet.update({
       where: {
@@ -77,6 +130,10 @@ export async function editPet(petId: unknown, newPetData: unknown) {
 export async function deletePet(petId: unknown) {
   await sleep(1000);
 
+  //authentication check -> i.e their token is valid
+  const session = await checkAuth();
+
+  //validate petId
   const validatedPetId = petIdSchema.safeParse(petId);
 
   if (!validatedPetId.success) {
@@ -85,6 +142,22 @@ export async function deletePet(petId: unknown) {
     };
   }
 
+  //authorization check -> i.e user is authorized to delete this pet
+  const pet = await getPetById(validatedPetId.data);
+
+  if (!pet) {
+    return {
+      message: "Pet not found",
+    };
+  }
+
+  if (pet.userId !== session.user.id) {
+    return {
+      message: "Unauthorized",
+    };
+  }
+
+  //database mutation
   try {
     await prisma.pet.delete({
       where: {
